Fix type of WebhookEvent.workflow_run_id

+++ gitspatch/models/_webhook.py
@@ -1,5 +1,4 @@
-from sqlalchemy import ForeignKey, String
-from sqlalchemy.dialects.postgresql import BIGINT
+from sqlalchemy import BigInteger, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from gitspatch.models._timestamp import TimestampMixin
@@ -19,9 +18,9 @@ class Webhook(IDModel, TimestampMixin, Base):
     )
     user: Mapped[User] = relationship("User", lazy="raise")
 
-    repository_id: Mapped[int] = mapped_column(BIGINT, nullable=False)
+    repository_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     workflow_id: Mapped[str] = mapped_column(String, nullable=False)
-    installation_id: Mapped[int] = mapped_column(BIGINT, nullable=False)
+    installation_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
 
     owner: Mapped[str] = mapped_column(String, nullable=False)
     repository_name: Mapped[str] = mapped_column(String, nullable=False)
